fix:타입 수정

+++ NWO_RegionNode/User.cs
@@ -8,9 +8,9 @@ public class User
     public int id = 0;
     public Vector3 position = new Vector3(0, 0, 0);
     public int speed = 0;
-    public int rot = 0;
+    public byte rot = 0;
 
-    public User(IChannelHandlerContext context, int id, Vector3 position,int spd, int rot)
+    public User(IChannelHandlerContext context, int id, Vector3 position,int spd,byte rot)
     {
         IChannel = context;
         this.id = id;
